Refactor electron metadata
